utils/testing: split module into a package
