+++ setup.cfg
@@ -46,6 +46,7 @@ packages =
 	plover.oslayer
 	plover.system
 	plover_build_utils
+	plover_build_utils.testing
 
 [options.entry_points]
 console_scripts =
